garden_academy/requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==19.9.0
# psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
djangorestframework == 3.9.2
djangorestframework_simplejwt
django-cors-headers
django-cors-middleware
psycopg2-binary